// web/src/shared/api/generated/model/createTeacherThreadBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { IFormFile } from './iFormFile';

export type CreateTeacherThreadBody = {
  text?: string;
} & {
  lessonId?: string;
} & {
  questionId?: string;
} & {
  attemptId?: string;
} & {
  image?: IFormFile;
};